% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{impute_column_values_pmm}
\alias{impute_column_values_pmm}
\title{k-NN Impute using Predictive Mean Matching}
\usage{
impute_column_values_pmm(
  result,
  obj,
  miss,
  col_offset,
  target_col_idx,
  nn_columns,
  nn_weights,
  n_imp,
  n_pmm,
  seed
)
}
\arguments{
\item{result}{Result matrix where column 2:2+n_imp contains imputed values (modified in place).}

\item{obj}{R matrix}

\item{miss}{is.na(obj).}

\item{col_offset}{Basically \code{cumsum(colSums(is.na(obj)))}.}

\item{target_col_idx}{The index of the target column to impute in obj.}

\item{nn_columns}{Column idx of nearest neighbors.}

\item{nn_weights}{Weights corresponding to nn_columns.}

\item{n_imp}{The number of imputations to perform for each missing value.}

\item{n_pmm}{The number of closest donors to consider for PMM.}

\item{seed}{RNG seed.}
}
\value{
void (modifies the result matrix in place).
}
\description{
Rcpp implementation of predictive mean matching (PMM) for imputation. Exported for testing only.
}
