% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{impute_knn_mlpack}
\alias{impute_knn_mlpack}
\title{Impute missing values in a matrix using treed k-nearest neighbors (k-NN)}
\usage{
impute_knn_mlpack(
  obj,
  miss,
  k,
  n_col_miss,
  method,
  tree,
  weighted,
  dist_pow,
  cores
)
}
\arguments{
\item{obj}{Numeric matrix with missing values pre-filled with colMeans.}

\item{miss}{Logical matrix (0/1) indicating missing values (1 = missing).}

\item{k}{Number of nearest neighbors to use for imputation.}

\item{n_col_miss}{Integer vector specifying the count of missing values per column.}

\item{method}{Integer specifying the distance metric: 0 = Euclidean, 1 = Manhattan.}

\item{tree}{Which type of tree? "kd" or "ball".}

\item{weighted}{Boolean controls for the imputed value to be a simple mean or weighted mean by inverse distance.}

\item{dist_pow}{A positive double that controls the penalty for larger distances in
the weighted mean imputation. Must be greater than zero: values between 0 and 1 apply a softer penalty,
1 is linear (default), and values greater than 1 apply a harsher penalty.}

\item{cores}{Number of CPU cores to use for parallel processing (default = 1).}
}
\value{
A matrix where the first column is the index of missing value as if calculated in R and column > 2 are imputed values.
}
\description{
k-NN using KDTree or BallTree.
}
