% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.ImputedMatrix}
\alias{print.ImputedMatrix}
\title{Print ImputedMatrix}
\usage{
\method{print}{ImputedMatrix}(x, n = 5, m = 5, ...)
}
\arguments{
\item{x}{An ImputedMatrix}

\item{n}{Number of rows to print}

\item{m}{Number of cols to print}

\item{...}{Not used}
}
\value{
Invisible object of class ImputedMatrix
}
\description{
Print ImputedMatrix
}
\examples{
data(khanmiss1)
t_khanmiss1 <- t(khanmiss1)
result <- knn_imp(t_khanmiss1, k = 5)
print(result, n = 6, m = 6)
}
