% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.KnnImpList}
\alias{print.KnnImpList}
\title{Print method for KnnImpList}
\usage{
\method{print}{KnnImpList}(x, m = 5, n = 10, ...)
}
\arguments{
\item{x}{A KnnImpList object}

\item{m}{Number of rows to show in preview (default: 5)}

\item{n}{Number of columns to show in preview (default: 10)}

\item{...}{Additional arguments (not used)}
}
\description{
Print method for KnnImpList
}
