% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.slide_impList}
\alias{print.slide_impList}
\title{Print method for slide_impList}
\usage{
\method{print}{slide_impList}(x, m = 5, n = 10, ...)
}
\arguments{
\item{x}{A slide_impList object}

\item{m}{Number of rows to show in preview (default: 5)}

\item{n}{Number of columns to show in preview (default: 10)}

\item{...}{Additional arguments (not used)}
}
\description{
Print method for slide_impList
}
