% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.slideimp_results}
\alias{print.slideimp_results}
\title{Print a \code{slideimp_results} Object}
\usage{
\method{print}{slideimp_results}(x, n = 6L, p = 6L, ...)
}
\arguments{
\item{x}{A \code{slideimp_results} object.}

\item{n}{Number of rows to print.}

\item{p}{Number of columns to print.}

\item{...}{Not used.}
}
\value{
\code{x}, invisibly.
}
\description{
Print the output of \code{\link[=knn_imp]{knn_imp()}}, \code{\link[=pca_imp]{pca_imp()}}, \code{\link[=group_imp]{group_imp()}}, or
\code{\link[=slide_imp]{slide_imp()}}.
}
\examples{
set.seed(1234)
mat <- sim_mat(n = 10, p = 10)
result <- knn_imp(mat$input, k = 5, .progress = FALSE)
class(result)
print(result, n = 6, p = 6)

}
