% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.slideimp_sim}
\alias{print.slideimp_sim}
\title{Print a \code{slideimp_sim} Object}
\usage{
\method{print}{slideimp_sim}(x, n = 6L, p = 6L, ...)
}
\arguments{
\item{x}{A \code{slideimp_sim} object.}

\item{n}{Number of rows of each component to show.}

\item{p}{Number of columns of \code{input} to show.}

\item{...}{Not used.}
}
\value{
\code{x}, invisibly.
}
\description{
Print the output of \code{\link[=sim_mat]{sim_mat()}}.
}
\examples{
set.seed(123)
sim_data <- sim_mat(n = 50, p = 10, rho = 0.5)
class(sim_data)
print(sim_data)

}
