% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.slideimp_tbl}
\alias{print.slideimp_tbl}
\title{Print a \code{slideimp_tbl} Object}
\usage{
\method{print}{slideimp_tbl}(x, n = NULL, ...)
}
\arguments{
\item{x}{A \code{slideimp_tbl} object.}

\item{n}{Number of rows to show. If \code{NULL}, a default is used.}

\item{...}{Not used.}
}
\value{
\code{x}, invisibly.
}
\description{
Print \code{slideimp_tbl} objects, which inherit from \code{data.frame}, with compact
display of list-columns.
}
\examples{
sim <- sim_mat(n = 10, p = 20)
tbl <- prep_groups(colnames(sim$input), sim$col_group)
class(tbl)
print(tbl)

}
