% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.SlideImpSimMat}
\alias{print.SlideImpSimMat}
\title{Print a SlideImpSimMat}
\usage{
\method{print}{SlideImpSimMat}(x, n = 6L, m = 6L, ...)
}
\arguments{
\item{x}{A \code{SlideImpSimMat} object.}

\item{n}{Number of rows of each component to show.}

\item{m}{Number of columns of \code{input} to show.}

\item{...}{Not used.}
}
\value{
Invisibly, \code{x}.
}
\description{
Print a SlideImpSimMat
}
