% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.SlideKnnList}
\alias{print.SlideKnnList}
\title{Print method for SlideKnnList}
\usage{
\method{print}{SlideKnnList}(x, m = 5, n = 10, ...)
}
\arguments{
\item{x}{A SlideKnnList object}

\item{m}{Number of rows to show in preview (default: 5)}

\item{n}{Number of columns to show in preview (default: 10)}

\item{...}{Additional arguments (not used)}
}
\description{
Print method for SlideKnnList
}
