% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SlideKnn.R
\name{restore_dimnames}
\alias{restore_dimnames}
\alias{restore_dimnames.SlideKnnList}
\alias{restore_dimnames.KnnImpList}
\title{Restore Dimnames of \code{\link[=SlideKnn]{SlideKnn()}}/\code{\link[=knn_imp]{knn_imp()}} Output}
\usage{
restore_dimnames(obj)

\method{restore_dimnames}{SlideKnnList}(obj)

\method{restore_dimnames}{KnnImpList}(obj)
}
\arguments{
\item{obj}{Output of \code{\link[=SlideKnn]{SlideKnn()}} or \code{\link[=knn_imp]{knn_imp()}}; a list with class
\code{SlideKnnList} or \code{KnnImpList}.}
}
\value{
The invisible input \code{obj} with dimnames restored if possible.
}
\description{
\code{\link[=SlideKnn]{SlideKnn()}}/\code{\link[=knn_imp]{knn_imp()}} can return a list of \code{\link[bigmemory:big.matrix]{bigmemory::big.matrix()}} objects
that will have dimnames stripped if \code{options(bigmemory.allow.dimnames = TRUE)}
is not set beforehand. This function attempts to restore the dimnames of each
object in the returned list using the "rownames" and "colnames" attributes
stored on the list.
}
\examples{
to_test <- t(
  sim_mat(
    n = 280,
    m = 100,
    perc_NA = 0.5,
    perc_col_NA = 1
  )$input
)
# manually strip dimnames from results
results <- knn_imp(to_test, k = 10, output = withr::local_tempfile())
on.exit(options(bigmemory.allow.dimnames = getOption("bigmemory.allow.dimnames")), add = TRUE)
options(bigmemory.allow.dimnames = TRUE)
rownames(results[[1]]) <- NULL
colnames(results[[1]]) <- NULL
# We see that dimnames have been stripped
results
# `restore_dimnames` stored in the object
restore_dimnames(results)
results
}
