% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group_imp.R
\name{slideimp_resolve_group}
\alias{slideimp_resolve_group}
\alias{slideimp_resolve_group.data.frame}
\alias{slideimp_resolve_group.default}
\title{Resolve a Group Specification to a Data Frame}
\usage{
slideimp_resolve_group(x)

\method{slideimp_resolve_group}{data.frame}(x)

\method{slideimp_resolve_group}{default}(x)
}
\arguments{
\item{x}{A group specification. \code{slideimp} provides a method for
\code{data.frame} objects. The optional \code{slideimp.extra} package provides a
method for character platform names.}
}
\description{
Convert a group specification to the canonical data-frame form expected by
\code{\link[=prep_groups]{prep_groups()}}. This S3 generic is exported so that extension packages,
such as \code{slideimp.extra}, can register additional methods.
}
\note{
This is primarily an extension hook for \code{slideimp.extra}.
}
\examples{
df <- data.frame(feature = c("cg1", "cg2"), group = c(1, 1))
slideimp_resolve_group(df)

}
