% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{weighted_row_means}
\alias{weighted_row_means}
\title{Weighted Row Mean}
\usage{
weighted_row_means(obj, miss, nn_columns, nn_weights)
}
\arguments{
\item{obj}{A numeric matrix containing the data.}

\item{miss}{An unsigned integer matrix indicating missing values (0 for observed, 1 for missing).}

\item{nn_columns}{An unsigned integer vector of column indices for the neighbors.}

\item{nn_weights}{A numeric vector of weights corresponding to the neighbors.}
}
\value{
A column vector containing the weighted row means, with NaN where computation is not possible.
}
\description{
Calculate weighted row means for specified columns, accounting for missing values.
}
