% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/col_vars.R
\name{col_vars}
\alias{col_vars}
\title{Calculate Matrix Column Variance}
\usage{
col_vars(mat, cores = 1)
}
\arguments{
\item{mat}{A numeric matrix.}

\item{cores}{Number of cores to use for parallel computation. Defaults to 1.}
}
\value{
A named numeric vector of column variances
}
\description{
Computes the sample variance for each column of a numeric matrix
}
\details{
Variances for columns with 1 value after dropping \code{NA} are set to \code{NA}.
}
\examples{
col_vars(t(khanmiss1))
}
