% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tune_imp.R
\name{compute_metrics}
\alias{compute_metrics}
\alias{compute_metrics.data.frame}
\alias{compute_metrics.slideimp_tune}
\title{Compute Prediction Accuracy Metrics}
\usage{
compute_metrics(results, metrics = c("mae", "rmse"))

\method{compute_metrics}{data.frame}(results, metrics = c("mae", "rmse"))

\method{compute_metrics}{slideimp_tune}(results, metrics = c("mae", "rmse"))
}
\arguments{
\item{results}{A \code{slideimp_tune} data.frame from \code{\link[=tune_imp]{tune_imp()}} containing
a \code{result} list-column with data.frames that have \code{truth} and \code{estimate} columns.}

\item{metrics}{A character vector of metric names to compute. Defaults
to \code{c("mae", "rmse")}. Also available: \code{"mape"}, \code{"bias"}, \code{"rsq"}, and \code{"rsq_trad"}.}
}
\value{
A data.frame with the original parameters and un-nested metrics
(\code{.metric}, \code{.estimator}, \code{.estimate}).
}
\description{
Computes prediction accuracy metrics for results from \code{\link[=tune_imp]{tune_imp()}}.
}
\details{
For alternative or faster metrics, see the \code{{yardstick}} package.
}
\examples{
data(khanmiss1)

set.seed(1234)
results <- tune_imp(
  obj = t(khanmiss1),
  parameters = data.frame(k = 10),
  .f = "knn_imp",
  n_reps = 1,
  num_na = 20
)

compute_metrics(results)

}
