% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SlideKnn.R
\name{mean_impute_row}
\alias{mean_impute_row}
\title{Row Mean Imputation}
\usage{
mean_impute_row(obj)
}
\arguments{
\item{obj}{A numeric matrix with \strong{samples in rows} and \strong{features in columns}. See details.
Can also be a path to the description file of, or a \code{\link[bigmemory:big.matrix]{bigmemory::big.matrix()}}.}
}
\value{
The matrix with NA values replaced by row means.
}
\description{
Imputes missing values (NA) in a matrix by replacing them with the mean of their
respective rows.
}
