% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.slideimp_results}
\alias{print.slideimp_results}
\title{Print a \code{slideimp_results} Object}
\usage{
\method{print}{slideimp_results}(x, n = 6L, p = 6L, ...)
}
\arguments{
\item{x}{An \code{slideimp_results} object}

\item{n}{Number of rows to print}

\item{p}{Number of cols to print}

\item{...}{Not used}
}
\value{
Invisible \code{x}
}
\description{
Print a \code{slideimp_results} Object
}
\examples{
data(khanmiss1)
t_khanmiss1 <- t(khanmiss1)
result <- knn_imp(t_khanmiss1, k = 5)
class(result)
print(result, n = 6, p = 6)
}
