% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.slideimp_tbl}
\alias{print.slideimp_tbl}
\title{Print a \code{slideimp_tbl} Object}
\usage{
\method{print}{slideimp_tbl}(x, n = NULL, ...)
}
\arguments{
\item{x}{A \code{slideimp_tbl} object.}

\item{n}{Number of rows to show. Defaults to 10.}

\item{...}{Not used.}
}
\value{
Invisible \code{x}
}
\description{
Print a \code{slideimp_tbl} Object
}
\examples{
data(khanmiss1)
set.seed(1234)
results <- tune_imp(t(khanmiss1), parameters = data.frame(k = 5), .f = "knn_imp")
class(results)
print(results)
}
