% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_mat.R
\name{sim_mat}
\alias{sim_mat}
\title{Simulate Matrix with Metadata}
\usage{
sim_mat(
  n = 100,
  p = 100,
  rho = 0.5,
  n_col_groups = 2,
  n_row_groups = 1,
  perc_total_na = 0.1,
  perc_col_na = 0.5,
  beta = TRUE
)
}
\arguments{
\item{n}{An integer specifying the number of rows (samples). Default is \code{100}.}

\item{p}{An integer specifying the number of columns (features). Default is \code{100}.}

\item{rho}{Columns correlation before scaling. Default is \code{0.5}.}

\item{n_col_groups}{An integer for the number of groups to assign to features/columns. Default is \code{2}.}

\item{n_row_groups}{An integer for the number of groups to assign to samples/rows. Default is \code{1}.}

\item{perc_total_na}{Proportion of all cells to set to NA. Default is \code{0.1}.}

\item{perc_col_na}{Proportion of columns across which those NAs are spread. Default is \code{0.5}.}

\item{beta}{If TRUE (default) scale values between 0 and 1 column wise.}
}
\value{
An object of class \code{SlideImpSimMat} which is a list with \code{input},
\code{col_group}, and \code{row_group}.
}
\description{
Generates a matrix of random normal data then optionally scale to between 0
and 1 between columns. It also creates corresponding data frames for feature
(column) and sample (row) metadata and can optionally introduce \code{NA} values
into a specified proportion of rows. A correlation between columns \code{rho}
(before scaling) can be added.
}
\examples{
set.seed(123)
sim_data <- sim_mat(n = 50, p = 10, rho = 0.5)
sim_data
}
