% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slide_imp.R
\name{slide_imp}
\alias{slide_imp}
\title{Sliding Window k-NN or PCA Imputation}
\usage{
slide_imp(
  obj,
  n_feat,
  n_overlap,
  k = NULL,
  colmax = 0.9,
  knn_method = c("euclidean", "manhattan"),
  cores = 1,
  post_imp = TRUE,
  dist_pow = 0,
  subset = NULL,
  ncp = NULL,
  scale = TRUE,
  pca_method = c("regularized", "em"),
  coeff.ridge = 1,
  seed = NULL,
  nb.init = 1,
  maxiter = 1000,
  miniter = 5,
  .progress = FALSE
)
}
\arguments{
\item{obj}{A numeric matrix with \strong{samples in rows} and \strong{features in columns}.}

\item{n_feat}{Number of features in a window.}

\item{n_overlap}{Number of overlapping features between two windows.}

\item{k}{Number of nearest neighbors for imputation. 10 is a good starting point.}

\item{colmax}{A number from 0 to 1. Threshold of missing data above which k-NN imputation is skipped.}

\item{knn_method}{Either "euclidean" (default) or "manhattan". Distance metric for nearest neighbor calculation.}

\item{cores}{Number of cores to parallelize over.}

\item{post_imp}{Logical flag indicating whether to impute remaining missing values (those that failed k-NN imputation) using column means.}

\item{dist_pow}{A numeric value controlling the degree of penalization of far-away nearest neighbors in the calculation of imputed values.
If \code{dist_pow = 0} (default), then the imputed value is just the mean of nearest neighbors regardless of distance.}

\item{subset}{Character vector of column names or integer vector of column
indices specifying which columns to impute.}

\item{ncp}{integer corresponding to the number of components used to to predict the missing entries}

\item{scale}{boolean. By default TRUE leading to a same weight for each variable}

\item{pca_method}{"Regularized" by default or "EM".}

\item{coeff.ridge}{1 by default to perform the regularized pca_imp (imputePCA) algorithm; useful only if method="Regularized". Other regularization terms can be implemented by setting the value to less than 1 in order to regularized less (to get closer to the results of the EM method) or more than 1 to regularized more (to get closer to the results of the mean imputation)}

\item{seed}{integer, by default seed = NULL implies that missing values are initially imputed by the mean of each variable. Other values leads to a random initialization}

\item{nb.init}{integer corresponding to the number of random initializations; the first initialization is the initialization with the mean imputation}

\item{maxiter}{integer, maximum number of iteration for the algorithm}

\item{miniter}{integer, minimum number of iteration for the algorithm}

\item{.progress}{Show progress bar (default = FALSE).}
}
\description{
Performs k-NN or PCA imputation large numeric matrices using a sliding
window approach column-wise. This method assumes that columns are meaningfully sorted.
}
\details{
The sliding window approach divides the input matrix into smaller, overlapping
segments and applies imputation to each window independently. Values in overlapping
areas are averaged across windows to produce the final imputed result.
This approach assumes that features (columns) are sorted meaningfully (e.g.,
by genomic position, time series, etc.).

Specify \code{k} and related arguments to use k-NN, \code{ncp} and related arguments for PCA.
}
\examples{
# Generate sample data with missing values with 20 samples and 100 columns
# where the column order is sorted (e.g., by genomic position or time)

set.seed(1234)
beta_matrix <- t(sim_mat(100, 20)$input)

# Sliding Window k-NN imputation by specifying `k`
imputed_knn <- slide_imp(
  beta_matrix,
  k = 5,
  n_feat = 50,
  n_overlap = 10
)
imputed_knn

# Sliding Window PCA imputation by specifying `ncp`
pca_knn <- slide_imp(
  beta_matrix,
  ncp = 2,
  n_feat = 50,
  n_overlap = 10
)
pca_knn

}
