% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slideimp-package.R
\docType{package}
\name{slideimp-package}
\alias{slideimp}
\alias{slideimp-package}
\title{slideimp: Numeric Matrices K-NN and PCA Imputation}
\description{
Provides K-nearest neighbor, PCA, grouped, and sliding-window imputation
methods for numeric matrices, with utilities for simulation and parameter
tuning.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/hhp94/slideimp}
  \item \url{https://hhp94.github.io/slideimp/}
  \item Report bugs at \url{https://github.com/hhp94/slideimp/issues}
}

}
\author{
\strong{Maintainer}: Hung Pham \email{amser.hoanghung@gmail.com} (\href{https://orcid.org/0000-0002-8271-9355}{ORCID}) [copyright holder]

}
\keyword{internal}
